requirements.txt
flask
numpy
opencv-python-headless
torch==1.9.1+cpu -f https://download.pytorch.org/whl/torch_stable.html
Pillow
requests
gunicorn